feat: Use rational number for progress reporting

+++ src/box360/_terraform.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <je2be/integers.hpp>
 #include <je2be/status.hpp>
 
 #include <minecraft-file.hpp>
@@ -15,7 +16,7 @@ class Terraform {
   class Impl;
 
 public:
-  static Status Do(mcfile::Dimension dim, std::filesystem::path const &poiDirectory, std::filesystem::path const &directory, unsigned int concurrency, Progress *progress, int progressChunksOffset);
+  static Status Do(mcfile::Dimension dim, std::filesystem::path const &poiDirectory, std::filesystem::path const &directory, unsigned int concurrency, Progress *progress, u64 progressChunksOffset);
 };
 
 } // namespace je2be::box360
