+++ src/toje-context.cpp
@@ -114,7 +114,7 @@ public:
                                        Options opt,
                                        mcfile::Endian endian,
                                        std::map<mcfile::Dimension, std::vector<std::pair<Pos2i, ChunksInRegion>>> &regions,
-                                       int &totalChunks,
+                                       u64 &totalChunks,
                                        i64 gameTick,
                                        GameMode gameMode,
                                        unsigned int concurrency) {
@@ -231,7 +231,7 @@ std::shared_ptr<Context> Context::Init(std::filesystem::path const &dbname,
                                        Options opt,
                                        mcfile::Endian endian,
                                        std::map<mcfile::Dimension, std::vector<std::pair<Pos2i, ChunksInRegion>>> &regions,
-                                       int &totalChunks,
+                                       u64 &totalChunks,
                                        i64 gameTick,
                                        GameMode gameMode,
                                        unsigned int concurrency) {
